+++ ft_printf/ft_puthex.c
@@ -1,20 +1,32 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_puthex.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ykuo <marvin@42.fr>                        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/10 19:01:55 by ykuo              #+#    #+#             */
+/*   Updated: 2022/05/10 19:02:42 by ykuo             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "ft_printf.h"
 
-void	ft_puthex (unsigned int n, char *base, int *len)
+void	ft_puthex(unsigned int n, char *base, int *len)
 {
-	if (n >15)
+	if (n > 15)
 		ft_puthex(n / 16, base, len);
 	ft_putchar(base[n % 16], len);
 }
 
-void	ft_puthex_pointer (unsigned long long n, char *base, int *len)
+void	ft_puthex_pointer(unsigned long n, char *base, int *len)
 {
-	if (n >15)
+	if (n > 15)
 		ft_puthex_pointer (n / 16, base, len);
 	ft_putchar(base[n % 16], len);
 }
 
-void	func_p (unsigned long long n, char *base, int *len)
+void	func_p(unsigned long n, char *base, int *len)
 {
 	if (n == 0)
 		ft_putstr ("(nil)", len);
